Fixed UB on OS X
Since - for some reason - OS X needs the red color component to be shifted by 24 to the left, and `int` is 32 bits wide on OS X, using any color of which the red component is greater than 127 results in shifting a `1` to the sign bit of a signed `int`, which results in undefined behavior. This commit fixes that by declaring `draw_clr` as an `unsigned` integer.